Display scores in info-box
