+++ app/lib/frontend/templates/listing.dart
@@ -75,11 +75,7 @@ String renderPackageList(
         package: view,
         searchQuery: searchQuery,
       ),
-      'like_score_html': _renderLabeledScore('likes', view.likes, ''),
-      'popularity_score_html':
-          _renderLabeledScore('popularity', view.popularity, '%'),
-      'health_score_html':
-          _renderLabeledScore('health', view.grantedPubPoints, ''),
+      'labeled_scores_html': renderLabeledScores(view),
       'has_api_pages': view.apiPages != null && view.apiPages.isNotEmpty,
       'api_pages': view.apiPages
           ?.map((page) => {
@@ -96,16 +92,6 @@ String renderPackageList(
   }
   return templateCache.renderTemplate('pkg/package_list', {
     'packages': packagesJson,
-  });
-}
-
-/// Renders the `views/pkg/labeled_score.mustache` template.
-String _renderLabeledScore(String label, int value, String sign) {
-  return templateCache.renderTemplate('pkg/labeled_score', {
-    'label': label,
-    'has_value': value != null,
-    'value': value ?? '--',
-    'sign': sign,
   });
 }
