+++ app/lib/frontend/templates/views/pkg/analysis/tab_experimental.mustache
@@ -4,8 +4,8 @@
 
 <div class="score-key-figures">
   {{& like_key_figure_html }}
-  {{& popularity_key_figure_html }}
   {{& pubpoints_key_figure_html }}
+  {{& popularity_key_figure_html }}
 </div>
 
 {{& report_html }}
