+++ app/lib/frontend/templates/views/pkg/info_box_experimental.mustache
@@ -2,6 +2,8 @@
     for details. All rights reserved. Use of this source code is governed by a
     BSD-style license that can be found in the LICENSE file. }}
 
+{{& labeled_scores_html }}
+
 {{#publisher_id}}
 <h3 class="title">Publisher</h3>
 <p>
