+++ app/lib/frontend/templates/views/pkg/labeled_scores.mustache
@@ -1,0 +1,15 @@
+{{! Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
+    for details. All rights reserved. Use of this source code is governed by a
+    BSD-style license that can be found in the LICENSE file. }}
+
+<div class="packages-scores">
+  <div class="packages-score packages-score-like">
+    {{& like_score_html }}
+  </div>
+  <div class="packages-score packages-score-health">
+    {{& pub_points_html }}
+  </div>
+  <div class="packages-score packages-score-popularity">
+    {{& popularity_score_html }}
+  </div>
+</div>
