+++ pkg/web_css/lib/src/_detail_page_experimental.scss
@@ -357,27 +357,6 @@ $detail-tabs-tablet-width: calc(100% - 240px);
   > .tab-button.-active[data-name="-admin-tab-"] {
     border-bottom-color: $detail-tab-admin-color;
   }
-
-  .detail-tab-analysis-title {
-    // Separate the tab title's bottom border from the score.
-    position: relative;
-    margin-right: 60px !important; // prevent override from :last-child
-    > a {
-      margin-right: 0px;
-    }
-  }
-
-  .score-value {
-    background: $detail-tab-active-fg;
-    color: white;
-    display: inline-block;
-    font-weight: 400;
-    padding: 0px 8px;
-    border-radius: 12px;
-    // Separate the tab title's bottom border from the score.
-    position: absolute;
-    margin-left: 8px;
-  }
 }
 
 .detail-tabs-content {
