+++ pkg/web_css/lib/src/_list_experimental.scss
@@ -292,72 +292,6 @@ body.experimental {
     }
   }
 
-  .packages-scores {
-    display: flex;
-    align-items: center;
-
-    font-family: $font-family-google-sans;
-
-    .packages-score {
-      min-width: 45px;
-
-      @media (min-width: $device-desktop-min-width) {
-        min-width: 55px;
-      }
-    }
-
-    .packages-score:not(:last-child) {
-      border-right: 1px solid #e0e0e0;
-      margin-right: 4px;
-      padding-right: 4px;
-
-      @media (min-width: $device-desktop-min-width) {
-        margin-right: 8px;
-        padding-right: 8px;
-      }
-    }
-
-    .packages-score-value {
-      color: #6d7278;
-      font-weight: 400;
-      text-align: center;
-
-      &.-has-value {
-        color: $color-input-primary;
-      }
-    }
-
-    .packages-score-value-number {
-      font-size: 18px;
-
-      @media (min-width: $device-desktop-min-width) {
-        font-size: 22px;
-      }
-    }
-
-    .packages-score-value-sign {
-      font-size: 10px;
-      vertical-align: super;
-
-      @media (min-width: $device-desktop-min-width) {
-        font-size: 12px;
-      }
-    }
-
-    .packages-score-label {
-      color: #6d7278;
-      font-size: 7px;
-      letter-spacing: 0.1px;
-      text-transform: uppercase;
-      text-align: center;
-
-      @media (min-width: $device-desktop-min-width) {
-        font-size: 8px;
-        letter-spacing: 0.2px;
-      }
-    }
-  }
-
   .packages-description {
     margin: 4px 0px;
   }
