--- pkg/web_css/lib/src/_score_circle.scss
@@ -1,47 +1,0 @@
-/* Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
-   for details. All rights reserved. Use of this source code is governed by a
-   BSD-style license that can be found in the LICENSE file. */
-
-.score-circle {
-  width: 40px;
-  height: 40px;
-  background: white;
-  border-radius: 50%;
-  position: relative;
-
-  .score-circle-outline,
-  .score-circle-arc,
-  .score-circle-label {
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    border-radius: 50%;
-  }
-
-  .score-circle-arc {
-    background: $score-circle-outline;
-    stroke: $score-circle-stroke;
-  }
-
-  .score-circle-label {
-    margin: 10%;
-    background: white;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .score-circle-primary-label {
-    color: $score-circle-label;
-    font-size: 16px;
-    font-weight: 500;
-  }
-
-  .score-circle-secondary-label {
-    /* Uses the defaults. */
-  }
-}
