+++ pkg/web_css/lib/style.scss
@@ -12,6 +12,6 @@
 @import 'src/_home';
 @import 'src/_pkg';
 @import 'src/_pkg_experimental';
-@import 'src/_score_circle';
+@import 'src/_scores';
 @import 'src/_search';
 @import 'src/_tags';
